WIP - finish initial updates to apartments script
- move AddressCache to geocode_missing
- move dt_cols_to_str from bike parking code to shared utility
- add more output columns to display file
- other small fixes
